Revert "lib/framework: reorganize sink/utils"
