+++ src/sinks/clickhouse/sink.rs
@@ -3,7 +3,7 @@ use codecs::encoding::Transformer;
 use event::Event;
 use framework::HealthcheckError;
 use framework::http::{HttpClient, HttpError};
-use framework::sink::http::{HttpEventEncoder, HttpRetryLogic, HttpSink};
+use framework::sink::util::http::{HttpEventEncoder, HttpRetryLogic, HttpSink};
 use framework::sink::util::retries::{RetryAction, RetryLogic};
 use http::header::{CONTENT_ENCODING, CONTENT_TYPE};
 use http::{Request, StatusCode, Uri};
