+++ src/sinks/elasticsearch/service.rs
@@ -6,8 +6,8 @@ use bytes::Bytes;
 use bytesize::ByteSizeOf;
 use event::{EventFinalizers, EventStatus, Finalizable};
 use framework::http::{Auth, HttpClient};
-use framework::sink::http::HttpBatchService;
 use framework::sink::util::Compression;
+use framework::sink::util::http::HttpBatchService;
 use framework::sink::util::service::RequestConfig;
 use framework::stream::DriverResponse;
 use futures::future::BoxFuture;
