+++ src/sinks/http/config.rs
@@ -5,8 +5,8 @@ use configurable::configurable_component;
 use framework::batch::{BatchConfig, RealtimeSizeBasedDefaultBatchSettings};
 use framework::config::{DataType, SinkConfig, SinkContext, serde_http_method, serde_uri};
 use framework::http::{Auth, HttpClient};
-use framework::sink::http::{HttpService, http_response_retry_logic};
 use framework::sink::util::Compression;
+use framework::sink::util::http::{HttpService, http_response_retry_logic};
 use framework::sink::util::service::{RequestConfig, ServiceBuilderExt};
 use framework::tls::TlsConfig;
 use framework::{Healthcheck, HealthcheckError, Sink};
