+++ src/sinks/http/sink.rs
@@ -3,8 +3,8 @@ use std::fmt::Debug;
 use async_trait::async_trait;
 use event::{EventContainer, Events};
 use framework::StreamSink;
-use framework::sink::http::HttpRequest;
 use framework::sink::util::builder::SinkBuilderExt;
+use framework::sink::util::http::HttpRequest;
 use framework::stream::{BatcherSettings, DriverResponse};
 use futures::{StreamExt, stream::BoxStream};
 use tower::Service;
