+++ src/sinks/influxdb/config.rs
@@ -4,8 +4,8 @@ use configurable::configurable_component;
 use framework::batch::{BatchConfig, SinkBatchSettings};
 use framework::config::{DataType, SinkConfig, SinkContext};
 use framework::http::HttpClient;
-use framework::sink::http::{HttpService, http_response_retry_logic};
 use framework::sink::util::Compression;
+use framework::sink::util::http::{HttpService, http_response_retry_logic};
 use framework::sink::util::service::{RequestConfig, ServiceBuilderExt};
 use framework::template::Template;
 use framework::tls::TlsConfig;
