+++ src/sinks/influxdb/sink.rs
@@ -3,9 +3,9 @@ use std::fmt::Debug;
 use async_trait::async_trait;
 use event::Events;
 use framework::StreamSink;
-use framework::sink::http::HttpRequest;
 use framework::sink::util::Compression;
 use framework::sink::util::builder::SinkBuilderExt;
+use framework::sink::util::http::HttpRequest;
 use framework::stream::{BatcherSettings, DriverResponse};
 use framework::template::Template;
 use futures::{StreamExt, stream::BoxStream};
